res/requirements.txt
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
fonttools==4.55.3
imageio==2.36.1
imageio-ffmpeg==0.5.1
kiwisolver==1.4.8
matplotlib==3.10.0
moviepy==2.1.1
numpy==2.2.1
packaging==24.2
pillow==10.4.0
proglog==0.1.10
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
scipy==1.15.0
six==1.17.0
tqdm==4.67.1
